Refactor tools drawer
- add owner tab
- clearer sections
- improve spacing and copy
