+++ src/components/shared/MinimalCollapse.tsx
@@ -1,4 +1,4 @@
-import { Collapse } from 'antd'
+import { Collapse, CollapseProps } from 'antd'
 import CollapsePanel from 'antd/lib/collapse/CollapsePanel'
 import { CSSProperties, PropsWithChildren } from 'react'
 
@@ -6,7 +6,9 @@ export function MinimalCollapse({
   children,
   header,
   style,
-}: PropsWithChildren<{ header: string | JSX.Element; style?: CSSProperties }>) {
+  ...props
+}: PropsWithChildren<{ header: string | JSX.Element; style?: CSSProperties }> &
+  CollapseProps) {
   return (
     <Collapse
       bordered={false}
@@ -16,6 +18,7 @@ export function MinimalCollapse({
         ...style,
       }}
       className="minimal ant-collapse-header-text-primary ant-collapse-header-p-0"
+      {...props}
     >
       <CollapsePanel key="1" header={header}>
         {children}
